rename TypeF into SerializableTypeF as better document
